+++ main/src/mill/main/Resolve.scala
@@ -8,228 +8,6 @@ import mill.main.ResolveMetadata.singleModuleMeta
 
 import scala.reflect.ClassTag
 
-object ResolveMetadata extends Resolve[String] {
-  def singleModuleMeta(obj: Module, discover: Discover[_], isRootModule: Boolean): Seq[String] = {
-    val modules = obj.millModuleDirectChildren.map(_.toString)
-    val targets =
-      obj
-        .millInternal
-        .reflectAll[NamedTask[_]]
-        .map(_.toString)
-    val commands =
-      for {
-        (cls, entryPoints) <- discover.value
-        if cls.isAssignableFrom(obj.getClass)
-        ep <- entryPoints
-      } yield
-        if (isRootModule) ep._2.name
-        else s"$obj.${ep._2.name}"
-
-    modules ++ targets ++ commands
-  }
-
-  def endResolveLabel(
-      obj: Module,
-      last: String,
-      discover: Discover[_],
-      rest: Seq[String]
-  ): Either[String, Seq[String]] = {
-    def direct = singleModuleMeta(obj, discover, obj.millModuleSegments.value.isEmpty)
-    last match {
-      case "__" =>
-        Right(
-          // Filter out our own module in
-          obj.millInternal.modules.flatMap(m => singleModuleMeta(m, discover, m == obj))
-        )
-      case "_" => Right(direct)
-      case _ =>
-        direct.find(_.split('.').last == last) match {
-          case None =>
-            Resolve.errorMsgLabel(direct, Seq(Segment.Label(last)), obj.millModuleSegments.value)
-          case Some(s) => Right(Seq(s))
-        }
-    }
-  }
-
-  def endResolveCross(
-      obj: Module,
-      last: List[String],
-      discover: Discover[_],
-      rest: Seq[String]
-  ): Either[String, List[String]] = {
-    obj match {
-      case c: Cross[Module] =>
-        last match {
-          case List("__") => Right(c.items.map(_._2.toString))
-          case items =>
-            c.items
-              .filter(_._1.length == items.length)
-              .filter(_._1.zip(last).forall { case (a, b) => b == "_" || a.toString == b })
-              .map(_._2.toString) match {
-              case Nil =>
-                Resolve.errorMsgCross(
-                  c.items.map(_._1.map(_.toString)),
-                  last,
-                  obj.millModuleSegments.value
-                )
-              case res => Right(res)
-            }
-
-        }
-      case _ =>
-        Left(
-          Resolve.unableToResolve(Segment.Cross(last), obj.millModuleSegments.value) +
-            Resolve.hintListLabel(obj.millModuleSegments.value)
-        )
-    }
-  }
-}
-
-object ResolveSegments extends Resolve[Segments] {
-
-  override def endResolveCross(
-      obj: Module,
-      last: List[String],
-      discover: Discover[_],
-      rest: Seq[String]
-  ): Either[String, Seq[Segments]] = {
-    obj match {
-      case c: Cross[Module] =>
-        last match {
-          case List("__") => Right(c.items.map(_._2.millModuleSegments))
-          case items =>
-            c.items
-              .filter(_._1.length == items.length)
-              .filter(_._1.zip(last).forall { case (a, b) => b == "_" || a.toString == b })
-              .map(_._2.millModuleSegments) match {
-              case Nil =>
-                Resolve.errorMsgCross(
-                  c.items.map(_._1.map(_.toString)),
-                  last,
-                  obj.millModuleSegments.value
-                )
-              case res => Right(res)
-            }
-        }
-      case _ =>
-        Left(
-          Resolve.unableToResolve(Segment.Cross(last), obj.millModuleSegments.value) +
-            Resolve.hintListLabel(obj.millModuleSegments.value)
-        )
-    }
-  }
-
-  def endResolveLabel(
-      obj: Module,
-      last: String,
-      discover: Discover[_],
-      rest: Seq[String]
-  ): Either[String, Seq[Segments]] = {
-    val target =
-      obj
-        .millInternal
-        .reflectSingle[Target[_]](last)
-        .map(t => Right(t.ctx.segments))
-
-    val command =
-      Resolve
-        .invokeCommand(obj, last, discover.asInstanceOf[Discover[Module]], rest)
-        .headOption
-        .map(_.map(_.ctx.segments))
-
-    val module =
-      obj.millInternal
-        .reflectNestedObjects[Module]
-        .find(_.millOuterCtx.segment == Segment.Label(last))
-        .map(m => Right(m.millModuleSegments))
-
-    command orElse target orElse module match {
-      case None =>
-        Resolve.errorMsgLabel(
-          singleModuleMeta(obj, discover, obj.millModuleSegments.value.isEmpty),
-          Seq(Segment.Label(last)),
-          obj.millModuleSegments.value
-        )
-
-      case Some(either) => either.right.map(Seq(_))
-    }
-  }
-}
-
-object ResolveTasks extends Resolve[NamedTask[Any]] {
-
-  def endResolveCross(
-      obj: Module,
-      last: List[String],
-      discover: Discover[_],
-      rest: Seq[String]
-  ): Either[String, Seq[NamedTask[Any]]] = {
-    obj match {
-      case c: Cross[Module] =>
-        Resolve.runDefault(obj, Segment.Cross(last), discover, rest).flatten.headOption match {
-          case None =>
-            Left(
-              "Cannot find default task to evaluate for module " +
-                Segments((obj.millModuleSegments.value :+ Segment.Cross(last)): _*).render
-            )
-          case Some(v) => v.map(Seq(_))
-        }
-      case _ =>
-        Left(
-          Resolve.unableToResolve(Segment.Cross(last), obj.millModuleSegments.value) +
-            Resolve.hintListLabel(obj.millModuleSegments.value)
-        )
-    }
-  }
-
-  def endResolveLabel(
-      obj: Module,
-      last: String,
-      discover: Discover[_],
-      rest: Seq[String]
-  ): Either[String, Seq[NamedTask[Any]]] = last match {
-    case "__" =>
-      Right(
-        obj.millInternal.modules
-          .filter(_ != obj)
-          .flatMap(m => m.millInternal.reflectAll[NamedTask[_]])
-      )
-    case "_" => Right(obj.millInternal.reflectAll[NamedTask[_]])
-
-    case _ =>
-      val target =
-        obj
-          .millInternal
-          .reflectSingle[NamedTask[_]](last)
-          .map(Right(_))
-
-      val command = Resolve.invokeCommand(
-        obj,
-        last,
-        discover.asInstanceOf[Discover[Module]],
-        rest
-      ).headOption
-
-      command orElse target orElse Resolve.runDefault(
-        obj,
-        Segment.Label(last),
-        discover,
-        rest
-      ).flatten.headOption match {
-        case None =>
-          Resolve.errorMsgLabel(
-            singleModuleMeta(obj, discover, obj.millModuleSegments.value.isEmpty),
-            Seq(Segment.Label(last)),
-            obj.millModuleSegments.value
-          )
-
-        // Contents of `either` *must* be a `Task`, because we only select
-        // methods returning `Task` in the discovery process
-        case Some(either) => either.map(Seq(_))
-      }
-  }
-}
-
 object Resolve {
   def minimum(i1: Int, i2: Int, i3: Int) = math.min(math.min(i1, i2), i3)
